+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Cameras/CameraViewPort.cpp
@@ -32,8 +32,8 @@ CameraViewPort::CameraViewPort()
 {
 }
 
-void	CameraViewPort::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	CameraViewPort::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 16)
 	{
