+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Line.cpp
@@ -35,7 +35,7 @@ Line::Line()
 {
 }
 
-void	Line::Process( uint32_t inStartOffset, uint32_t inEndOffset )
+void	Line::Process( size_t inStartOffset, size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 24)
 	{
