+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Marker.cpp
@@ -32,10 +32,10 @@ Marker::Marker()
 {
 }
 
-void	Marker::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Marker::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen < 36)
 	{
