+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Point.cpp
@@ -32,7 +32,7 @@ Point::Point()
 {
 }
 
-void	Point::Process( uint32_t inStartOffset, uint32_t inEndOffset )
+void	Point::Process( size_t inStartOffset, size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 12)
 	{
