+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Geometries/Point.h
@@ -30,6 +30,6 @@ class Point : public TypeHandler
 public:
 							Point();
 	
-	virtual void			Process( uint32_t inStartOffset,
-									uint32_t inEndOffset );
+	virtual void			Process( size_t inStartOffset,
+									size_t inEndOffset );
 };
