+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Miscellaneous/Reference.cpp
@@ -32,8 +32,8 @@ Reference::Reference()
 {
 }
 
-void	Reference::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	Reference::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset - inStartOffset != 4)
 	{
