+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Shaders/PixmapTexture.cpp
@@ -53,10 +53,10 @@ PixmapTexture::PixmapTexture()
 {
 }
 
-void	PixmapTexture::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	PixmapTexture::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen < 28)
 	{
