+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/InterpolationStyle.h
@@ -32,6 +32,6 @@ class InterpolationStyle : public TypeHandler
 public:
 							InterpolationStyle();
 	
-	virtual void			Process( uint32_t inStartOffset,
-									uint32_t inEndOffset );
+	virtual void			Process( size_t inStartOffset,
+									size_t inEndOffset );
 };
