+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Transforms/Matrix.cpp
@@ -34,7 +34,7 @@ Matrix::Matrix()
 {
 }
 
-void	Matrix::Process( uint32_t inStartOffset, uint32_t inEndOffset )
+void	Matrix::Process( size_t inStartOffset, size_t inEndOffset )
 {
 	
 	if (inEndOffset != inStartOffset + 64)
