+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Transforms/Rotate.cpp
@@ -37,8 +37,8 @@ Rotate::Rotate()
 {
 }
 
-void	Rotate::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Rotate::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 8)
 	{
