+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/View Objects/ImageClearColor.cpp
@@ -33,8 +33,8 @@ ImageClearColor::ImageClearColor()
 {
 }
 
-void	ImageClearColor::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	ImageClearColor::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 16)
 	{
